% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test.R
\name{f_test}
\alias{f_test}
\title{\eqn{F}-test}
\usage{
f_test(s_1, s_2, n_1, n_2, alfa = 0.05, dec = 10)
}
\arguments{
\item{s_1}{Smerodatna odchylka prvniho vyberu.}

\item{s_2}{Smerodatna odchylka druheho vyberu.}

\item{n_1}{Pocet pozorovani prvniho vyberu.}

\item{n_2}{Pocet pozorovani druheho vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium dvouvyberoveho \eqn{F}-testu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru \eqn{F}-testu.
}
