% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beta.R
\name{is_beta}
\alias{is_beta}
\title{Intervaly spolehlivosti pro regresni parametr \eqn{\beta}}
\usage{
is_beta(b, s, n, p, alfa = 0.05, dec = 10)
}
\arguments{
\item{b}{Odhad parametru beta.}

\item{s}{Smerodatna odchylka parametru beta.}

\item{n}{Pocet pozorovani.}

\item{p}{Pocet parametru v modelu (bez urovonove konstanty).}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro \eqn{\beta}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet instervalu spolehlivosti regresniho parametru \eqn{\beta}
}
