% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mu_s_nezname.R
\name{is_mu_s_nezname}
\alias{is_mu_s_nezname}
\title{Intervaly spolehlivosti pro \eqn{\mu}, kdyz \eqn{\sigma^2} nezname}
\usage{
is_mu_s_nezname(m, s, n, alfa = 0.05, dec = 10)
}
\arguments{
\item{m}{Stredni hodnota vyberu.}

\item{s}{Smerodatna odchylka vyberu.}

\item{n}{Pocet pozorovani.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro \eqn{\mu}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet instervalu spolehlivosti stredni hodnoty \eqn{\mu}, kdyz nezname hodnotu smerodatne odchylky \eqn{\sigma^2}.
}
