% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mu_s_nezname_2.R
\name{is_mu_s_nezname_2}
\alias{is_mu_s_nezname_2}
\title{Intervaly spolehlivosti pro \eqn{\mu_1 - \mu_2}, kdyz \eqn{latex}{\sigma_1, \sigma_2} nezname}
\usage{
is_mu_s_nezname_2(m_1, m_2, s_1, s_2, n_1, n_2, alfa = 0.05, dec = 10)
}
\arguments{
\item{m_1}{Stredni hodnota prvniho vyberu.}

\item{m_2}{Stredni hodnota druheho vyberu.}

\item{s_1}{Smerodatna odchylka prvniho vyberu.}

\item{s_2}{Smerodatna odchylka druheho vyberu.}

\item{n_1}{Pocet pozorovani v prvnim vyberu.}

\item{n_2}{Pocet pozorovani v prvnim vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro rozdil \eqn{\mu_1 - \mu_2}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet instervalu spolehlivosti rozdilu stredni hodnot \eqn{\mu_1 - \mu_2}, kdyz nezname hodnoty smerodatnych odchylek \eqn{\sigma_1, \sigma_2}, ale vime, ze jsou shodne.
}
