% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rho.R
\name{is_rho}
\alias{is_rho}
\title{Interval spolehlivosti pro korelacni koeficient \eqn{\rho}}
\usage{
is_rho(r, n, alfa = 0.05, dec = 10)
}
\arguments{
\item{n}{Pocet pozorovani.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}

\item{b}{Odhad parametru beta.}

\item{s}{Smerodatna odchylka parametru beta.}

\item{p}{Pocet parametru v modelu (bez urovonove konstanty).}
}
\value{
Vypocita oboustranny interval spolehlivosti pro \eqn{\beta}.
}
\description{
Funkce pro vypocet instervalu spolehlivosti korelacniho koeficientu \eqn{\rho}
}
