% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_s_mu_nezname.R
\name{is_s_mu_nezname}
\alias{is_s_mu_nezname}
\title{Intervaly spolehlivosti pro \eqn{\sigma^2}, kdyz \eqn{\mu} nezname}
\usage{
is_s_mu_nezname(s, n, alfa = 0.05, dec = 10)
}
\arguments{
\item{s}{Smerodatna odchylka vyberu.}

\item{n}{Pocet pozorovani.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro \eqn{\sigma^2}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet instervalu spolehlivosti rozptylu \eqn{\sigma^2}, kdyz nezname hodnotu stredni hodnoty \eqn{\mu}.
}
