% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_s_mu_nezname_2.R
\name{is_s_mu_nezname_2}
\alias{is_s_mu_nezname_2}
\title{Intervaly spolehlivosti pro spolecny neznamy rozptyl \eqn{\sigma^2}}
\usage{
is_s_mu_nezname_2(s_1, s_2, n_1, n_2, alfa = 0.05, dec = 10)
}
\arguments{
\item{s_1}{Smerodatna odchylka prvniho vyberu.}

\item{s_2}{Smerodatna odchylka druheho vyberu}

\item{n_1}{Pocet pozorovani prvniho vyberu.}

\item{n_2}{Pocet pozorovani druheho vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro spolecny rozptyl \eqn{\sigma^2}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet instervalu spolehlivosti spolecneho rozptylu \eqn{\sigma^2}.
}
