% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_s_podil_2.R
\name{is_s_podil_2}
\alias{is_s_podil_2}
\title{Intervaly spolehlivosti pro podil rozptylu \eqn{\sigma_1^2 / \sigma_2^2}}
\usage{
is_s_podil_2(s_1, s_2, n_1, n_2, alfa = 0.05, dec = 10)
}
\arguments{
\item{s_1}{Smerodatna odchylka prvniho vyberu.}

\item{s_2}{Smerodatna odchylka druheho vyberu.}

\item{n_1}{Pocet pozorovani v prvnim vyberu.}

\item{n_2}{Pocet pozorovani v prvnim vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro podil rozptylu \eqn{\sigma_1^2 / \sigma_2^2}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet intervalu spolehlivosti pro podil rozptylu \eqn{\sigma_1^2 / \sigma_2^2}.
}
