% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_theta.R
\name{is_theta}
\alias{is_theta}
\title{Intervaly spolehlivosti pro \eqn{\theta} z alternativniho rozlozeni}
\usage{
is_theta(m, n, alfa = 0.05, dec = 10)
}
\arguments{
\item{m}{Stredni hodnota podilu vyberu.}

\item{n}{Pocet pozorovani.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita intervaly spolehlivosti pro \eqn{\theta}, oboustranny, levo- a pravostranny.
}
\description{
Funkce pro vypocet intervalu spolehlivosti podilu \eqn{\theta}, pokud mame k dispozici data z alternativniho rozlozeni.
}
