% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_theta_2.R
\name{is_theta_2}
\alias{is_theta_2}
\title{Intervaly spolehlivosti pro rozdil parametru \eqn{\theta_1 - \theta_2}}
\usage{
is_theta_2(m_1, m_2, n_1, n_2, alfa = 0.05, dec = 10)
}
\arguments{
\item{m_1}{Stredni hodnota podilu prvniho vyberu.}

\item{m_2}{Stredni hodnota podilu druheho vyberu.}

\item{n_1}{Pocet pozorovani prvniho vyberu.}

\item{n_2}{Pocet pozorovani druheho vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium pro testovani parametru \eqn{\theta}.
}
\description{
Funkce pro vypocet instervalu spolehlivosti pro rozdil parametru \eqn{\theta_1 - \theta_2}.
}
