% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_test_0.R
\name{rho_test_0}
\alias{rho_test_0}
\title{\eqn{\rho}-test oproti 0}
\usage{
rho_test_0(r, n, alfa = 0.05, dec = 10)
}
\arguments{
\item{r}{Vyberovy korelacni koeficient.}

\item{n}{Pocet pozorovani.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium testu vyznamnosti korelacniho koeficientu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru pri testovani vyznamnosti korelacniho koeficientu (oproti nule).
}
