% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_test_2.R
\name{rho_test_2}
\alias{rho_test_2}
\title{Test o rovnosti dvou korelacnich koeficientu}
\usage{
rho_test_2(r_1, r_2, n_1, n_2, alfa = 0.05, dec = 10)
}
\arguments{
\item{r_1}{Vyberovy korelacni koeficient prvniho vyberu.}

\item{r_2}{Vyberovy korelacni koeficient druheho vyberu.}

\item{n_1}{Pocet pozorovani prvniho vyberu.}

\item{n_2}{Pocet pozorovani druheho vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium testu rovnosti korelacnich koeficientu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru pri testovani rovnosti dvou korelacnich koeficientu.
}
