% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_test_c.R
\name{rho_test_c}
\alias{rho_test_c}
\title{\eqn{\rho}-test oproti konstante}
\usage{
rho_test_c(r, n, c, alfa = 0.05, dec = 10)
}
\arguments{
\item{r}{Vyberovy korelacni koeficient.}

\item{n}{Pocet pozorovani.}

\item{c}{Konstanta, oproti ktere se testuje vyberovy korelacni koeficient.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium testu vyznamnosti korelacniho koeficientu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru pri testovani vyznamnosti korelacniho koeficientu (oproti konstante).
}
