% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozptyl_test.R
\name{rozptyl_test}
\alias{rozptyl_test}
\title{Test o rozptylu}
\usage{
rozptyl_test(s_2, n, c, alfa = 0.05, dec = 10)
}
\arguments{
\item{s_2}{Rozptyl vyberu.}

\item{n}{Pocet pozorovani.}

\item{c}{Konstanta, se kterou se porovnava rozptyl vyberu.}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium testu o rozptylu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru testu o rozptylu.
}
