% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_2.R
\name{t_test_2}
\alias{t_test_2}
\title{Dvouvyberovy \eqn{t}-test}
\usage{
t_test_2(m_1, m_2, s_1, s_2, n_1, n_2, c = 0, alfa = 0.05, dec = 10)
}
\arguments{
\item{m_1}{Stredni hodnota prvniho vyberu.}

\item{m_2}{Stredni hodnota druheho vyberu.}

\item{s_1}{Smerodatna odchylka prvniho vyberu.}

\item{s_2}{Smerodatna odchylka druheho vyberu.}

\item{n_1}{Pocet pozorovani prvniho vyberu.}

\item{n_2}{Pocet pozorovani druheho vyberu.}

\item{c}{Konstanta, se kterou se porovnava stredni hodnota vyberu (je 0, pokud neni stanoveho jinak).}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium dvouvyberoveho \eqn{t}-testu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru dvouvyberoveho \eqn{t}-testu.
}
