% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_beta.R
\name{t_test_beta}
\alias{t_test_beta}
\title{\eqn{t}-test beta}
\usage{
t_test_beta(b, s, n, p, alfa = 0.05, dec = 10)
}
\arguments{
\item{b}{Odhad parametru beta}

\item{s}{Smerodatna odchylka parametru beta.}

\item{n}{Pocet pozorovani.}

\item{p}{Pocet parametru v modelu (bez urovonove konstanty).}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium \eqn{t}-testu pro parametr beta.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru \eqn{t}-testu pro regresni parametr beta.
}
