% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_test_2.R
\name{theta_test_2}
\alias{theta_test_2}
\title{Testovani parametru \eqn{\theta} z alternativniho rozlozeni}
\usage{
theta_test_2(m_1, m_2, n_1, n_2, c = 0, alfa = 0.05, dec = 10)
}
\arguments{
\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}

\item{m}{Stredni hodnota podilu vyberu.}

\item{n}{Pocet pozorovani.}
}
\value{
Vypocita kriticke obory a testovaci kriterium pro testovani parametru \eqn{\theta}.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru rozdilu parametru \eqn{\theta_1 - \theta_2}.
}
