% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch.R
\name{welch}
\alias{welch}
\title{Welchova aproximace}
\usage{
welch(s_1, s_2, n_1, n_2, dec = 10)
}
\arguments{
\item{s_1}{Smerodatna odchylka prvniho vyberu.}

\item{s_2}{Smerodatna odchylka druheho vyberu.}

\item{n_1}{Pocet pozorovani v prvnim vyberu.}

\item{n_2}{Pocet pozorovani v prvnim vyberu.}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}
}
\value{
Provede vypocet Welchove aproximace (pokud vysledek neni cele cislo, doporucuje se linearni interpolace).
}
\description{
Funkce pro vypocet Welchovy aproximace, vysledek ktere pak muze byt pouzit jako stupne volnosti pro Studentovo rozlozeni.
}
