% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.R
\name{z_test}
\alias{z_test}
\title{\eqn{z}-test}
\usage{
z_test(m, s, n, c = 0, alfa = 0.05, dec = 10)
}
\arguments{
\item{m}{Stredni hodnota vyberu.}

\item{s}{Smerodatna odchylka.}

\item{n}{Pocet pozorovani.}

\item{c}{Konstanta, se kterou se porovnava stredni hodnota vyberu (je 0, pokud neni stanoveno jinak).}

\item{alfa}{Hladina vyznamnosti (je 0.05, pokud neni stanoveno jinak).}

\item{dec}{Pocet desetinnych mist v konecnem vysledku (je 10, pokud neni stanoveno jinak).}
}
\value{
Vypocita kriticke obory a testovaci kriterium \eqn{z}-testu.
}
\description{
Funkce pro vypocet testovaciho kriteria a kritickych oboru \eqn{z}-testu.
}
